- Implemented Kailipton ReturnToVillage skill (still working only in Laglamia map). - Added Kailipton ManaShield skill (not implemented Yet). - Updated some info from Item.dta file. - Improved use_skill and attack Server commands from PacketParser. - Fixed an issue detected when leveling up grouped skills. - Improved @resetskills command from MessageParser.
git-svn-id: http://reunion.googlecode.com/svn/trunk@218 d5b2e400-7aba-11de-9499-4f8beaa274e3

+++ jreunion/src/com/googlecode/reunion/jreunion/game/items/equipment/Weapon.java
@@ -34,7 +34,7 @@ public abstract class Weapon extends PlayerItem {
 
 	private int handed; // 1 - One handed; 2 - Two handed
 
-	private float magicDmg;
+	private float magicDmg; // value in %
 
 	public Weapon(int id) {
 		super(id);
@@ -184,7 +184,7 @@ public abstract class Weapon extends PlayerItem {
 			}	
 			if (item.checkMembers(new String[] { "MagicDmg" })) {
 				// use member from file
-				setMagicDmg(Integer.parseInt(item.getMemberValue("MagicDmg")));
+				setMagicDmg(Float.parseFloat(item.getMemberValue("MagicDmg")));
 			} else {
 				// use default
 				setMagicDmg(0);
@@ -236,7 +236,7 @@ public abstract class Weapon extends PlayerItem {
 		this.speed = speed;
 	}
 	
-	public void setMagicDmg(int magicDmg) {
+	public void setMagicDmg(float magicDmg) {
 		this.magicDmg = magicDmg;
 	}
